backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-genai>=1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
